+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## (unreleased) tdl@7.2.0
 
+- Added a `client.loginAsBot` function.
 - Added documentation to the `.d.ts` declaration file.
 
 ## tdl@7.1.0 (2022-09-18)